requirements/base.txt
Django==4.2.4
psycopg2==2.9.7
djangorestframework==3.14.0
